pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-server-template"
version = "0.1.0"
description = "A clean, simple MCP server template"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.1.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.100.0",
    "anyio>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
    "mypy",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov",
    "httpx>=0.24.0",
    "pre-commit",
    "pip-audit",
    "bandit",
]

[project.scripts]
mcp-server = "mcp_template.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_template"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "B", "UP"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
